# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a B2B-API server.",
        "title": "B2B Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/api/v1",
    "paths": {
        "/2fa/confirm-google-totp": {
            "get": {
                "description": "Google-Authenticator Confirmation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2FA"
                ],
                "summary": "Google-Authenticator Confirmation route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uint",
                        "description": " ",
                        "name": "otp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Token"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/2fa/send-sms-otp": {
            "get": {
                "description": "SMS sender",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2FA"
                ],
                "summary": "SMS sender route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Authentication route",
                "parameters": [
                    {
                        "format": "json",
                        "description": "authentication",
                        "name": "authentication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "get": {
                "description": "RefreshToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "RefreshToken route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "refreshToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Token"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorModel": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                },
                "response": {}
            }
        },
...
        "models.Login": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 9,
                    "minLength": 9
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                },
                "response": {
                    "$ref": "#/definitions/models.twoFa"
                }
            }
        },
        "models.OTP": {
            "type": "object",
            "properties": {
                "expirition": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "totalSent": {
                    "type": "integer"
                }
            }
        },
        "models.Token": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "models.googleAuth": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                }
            }
        },
        "models.smsOTP": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                }
            }
        },
        "models.twoFa": {
            "type": "object",
            "properties": {
                "googleAuthenticator": {
                    "$ref": "#/definitions/models.googleAuth"
                },
                "smsOtp": {
                    "$ref": "#/definitions/models.smsOTP"
                }
            }
        }
    }
}